add: view model to home page to follow MVVM
