+++ AshChatNative/src/app/private/home/index.tsx
@@ -7,24 +7,17 @@ import { Ionicons, MaterialIcons } from "@expo/vector-icons";
 import {useEffect, useState } from "react";
 import { Text, TouchableOpacity, View} from "react-native";
 import { useMMKVObject } from "react-native-mmkv";
+import { HomeViewModel } from "./home-view-model";
 
-export default function Home(){
-    const [isModalOpen, setIsModalOpen] = useState(false);
-    const [chatLabels] = useMMKVObject<LabelChatPropsDTO[]>("ashchat.label.chats")
-    const [typeOfLabelToShow, setTypeOfLabelToShow] = useState("all");
-    const [chatLabelsToShow, setChatLabelsToShow] = useState<LabelChatPropsDTO[] | undefined>(undefined);
-
-    function handleOpenModal() {
-        setIsModalOpen(!isModalOpen);
-    }
-    function filterChatsToShow(type: string) {
-        if (!chatLabels) return;
-        const filteredChats = type === "all" ? chatLabels : chatLabels.filter(chat => chat.notification > 0);
-        setChatLabelsToShow(filteredChats);
-    }
-    useEffect(() => {
-        filterChatsToShow(typeOfLabelToShow);
-    }, [chatLabels, typeOfLabelToShow]);
+export default function index(){
+    const {
+        chatLabelsToShow,
+        isModalOpen,
+        typeOfLabelToShow,
+        handleOpenModal,
+        handleSetTypeOfLabelToShow
+        
+    } = HomeViewModel();
     
     return (
         <View className="flex-1 pt-[62px] px-10" >
@@ -38,13 +31,13 @@ export default function Home(){
             <Text className="font-bold text-white text-3xl">My Chats</Text>
 
             <View className="flex-row items-center justify-between mt-5">
-                <TouchableOpacity className="flex-row items-center gap-2" onPress={() => setTypeOfLabelToShow("all")}>
+                <TouchableOpacity className="flex-row items-center gap-2" onPress={() => handleSetTypeOfLabelToShow("all")}>
                     <Ionicons name="chatbubble-outline" size={24} color="white" />
                     <Text className={`font-bold ${typeOfLabelToShow === "all" ? "text-purple-700" : "text-white"}`}>
                         All
                     </Text>
                 </TouchableOpacity>
-                <TouchableOpacity className="flex-row items-center gap-2" onPress={() => setTypeOfLabelToShow("unread")}>
+                <TouchableOpacity className="flex-row items-center gap-2" onPress={() => handleSetTypeOfLabelToShow("unread")}>
                     <Ionicons name="chatbubble-ellipses-outline" size={24} color="white" />
                     <Text className={`font-bold ${typeOfLabelToShow === "unread" ? "text-purple-700" : "text-white"}`}>
                         Unread
